+++ lang/en/theme_boost_union.php
@@ -321,6 +321,63 @@ $string['courseoverviewshowcourseimagessetting_desc'] = 'With this setting, you 
 $string['courseoverviewshowprogresssetting'] = 'Show course completion progress';
 $string['courseoverviewshowprogresssetting_desc'] = 'With this setting, you can control whether the course completion progress is visible inside the course overview block or not.';
 
+// Settings: Category index / site home tab.
+$string['categoryindextab'] = 'Category index / Site home';
+// ... Section: Course listing.
+$string['courselistingheading'] = 'Course listing';
+// ... ... Setting: Course listing presentation.
+$string['courselistingpresentation'] = 'Course listing presentation';
+$string['courselistingpresentation_desc'] = 'With this setting, you can modify the look & feel of the course listing on the category index pages and on site home. As an alternative to the way how Moodle core presents them, you can present the course listing as course cards (similar to the course cards on the \'My courses\' page) or as course list (similar to the course list on the \'My courses\' page).';
+$string['courselistingpresentation_nochange'] = 'Designer\'s nightmare (unchanged as presented by Moodle core)';
+$string['courselistingpresentation_cards'] = 'Course cards';
+$string['courselistingpresentation_list'] = 'Course list';
+$string['courselistingpresentation_note'] = 'Please note: If you enable course cards or course lists, the <a href="{$a->url1}">coursesperpage</a> setting is still respected and will control how many cards / rows will be shown. However, the <a href="{$a->url2}">courseswithsummarieslimit</a> setting does not have any effect anymore – all courses will be shown with full details. Please take care not to set <a href="{$a->url1}">coursesperpage</a> too high to avoid long page load times on pages with many courses.';
+// ... ... Setting: Course card column count.
+$string['coursecardscolumncount'] = 'Course card column count';
+$string['coursecardscolumncount_desc'] = 'The course card grid will be presented in a responsive way and its columns will wrap on smaller screens. With this setting, you just control the maximum number of columns in the course card grid on larger screens. Setting the maximum number of columns to 2 instead of 3 might make the course cards look more spacious and less crowded. Setting the maximum number of columns to 1 is possible as well and will effectively turn the course listing in a vertical list of cards.';
+// ... ... Setting: Show course image in the course listing.
+$string['courselistinghowimage'] = 'Show course image in the course listing';
+$string['courselistinghowimage_desc'] = 'With this setting, you control if the course image is shown in the course listing or not.';
+// ... ... Setting: Show course contacts in the course listing.
+$string['courselistingshowcontacts'] = 'Show course contacts in the course listing';
+$string['courselistingshowcontacts_desc'] = 'With this setting, you control if the course contact\'s pictures are shown in the course listing or not. Please note: The contact pictures are shown together with the course image, thus presenting course contacts without presenting the course image is not possible.';
+// ... ... Setting: Show course shortname in the course listing.
+$string['courselistinghowshortname'] = 'Show course shortname in the course listing';
+$string['courselistinghowshortname_desc'] = 'With this setting, you control if the course shortname is shown in the course listing or not.';
+// ... ... Setting: Show course category in the course listing.
+$string['courselistinghowcategory'] = 'Show course category in the course listing';
+$string['courselistinghowcategory_desc'] = 'With this setting, you control if the course category is shown in the course listing or not.';
+// ... ... Setting: Show course completion progress in the course listing.
+$string['courselistinghowprogress'] = 'Show course completion progress in the course listing';
+$string['courselistinghowprogress_desc'] = 'With this setting, you control if the course completion progress are shown in the course listing or not.';
+// ... ... Setting: Show course enrolment icons in the course listing.
+$string['courselistinghowenrolicons'] = 'Show course enrolment icons in the course listing';
+$string['courselistinghowenrolicons_desc'] = 'With this setting, you control if the course enrolment icons are shown in the course listing or not.';
+// ... ... Setting: Show course fields in the course listing.
+$string['courselistinghowfields'] = 'Show course fields in the course listing';
+$string['courselistinghowfields_desc'] = 'With this setting, you control if the custom course fields are shown in the course listing or not.';
+// ... ... Setting: Show goto button in the course listing.
+$string['courselistinghowgoto'] = 'Show goto button in the course listing';
+$string['courselistinghowgoto_desc'] = 'With this setting, you control if a \'Go to course\' button is shown in the course listing or not. If this setting is disabled, the user is still able to go to the course by clicking on the course title or course image.';
+$string['courselistinggoto'] = 'Go to course';
+// ... ... Setting: Show details popup in the course listing.
+$string['courselistinghowpopup'] = 'Show details popup in the course listing';
+$string['courselistinghowpopup_desc'] = 'With this setting, you control if a \'Course details\' button is shown in the course listing or not. With this button, the user can open a details popup which contains the course summary, the course contacts and the course fields. The popup will contain this information regardless if you enabled it on the course card / row itself or not.';
+$string['courselistingpopup'] = 'Details';
+$string['courselistingummary'] = 'Course summary';
+$string['courselistingnosummary'] = 'This course does not have a summary';
+$string['courselistingcontacts'] = 'Course contact';
+$string['courselistingviewprofile'] = 'View profile';
+$string['courselistingfields'] = 'Course classification';
+// ... Section: Category listing.
+$string['categorylistingheading'] = 'Category listing';
+// ... ... Setting: Category listing presentation.
+$string['categorylistingpresentation'] = 'Category listing presentation';
+$string['categorylistingpresentation_desc'] = 'With this setting, you can modify the look & feel of the category listing on the category index pages and on site home. As an alternative to the way how Moodle core presents them, you can present the category listing as a refreshed list of boxes.';
+$string['categorylistingpresentation_nochange'] = 'Designer\'s nightmare (unchanged as presented by Moodle core)';
+$string['categorylistingpresentation_boxlist'] = 'List of boxes';
+$string['categorylistingpresentation_note'] = 'Please note: If you enabled the \'Course listing presentation\' setting above, we recommend to enable this setting as well. Both were designed to work together.';
+
 // Settings: Blocks tab.
 // The string for this tab is the same as on the 'Feel' page.
 // ... Section: Timeline block.